# Single threaded pool buffer