--- a/internal/async/map_clear.go
+++ b/internal/async/map_clear.go
@@ -1,4 +1,4 @@
-//go:build !go1.23 && !migrated_fynedo
+//go:build !go1.23
 
 package async