--- a/internal/async/migrated/map.go
+++ b/internal/async/migrated/map.go
@@ -1,0 +1,10 @@
+//go:build !migrated_fynedo
+
+package migrated
+
+import "fyne.io/fyne/v2/internal/async"
+
+// Map is a generic wrapper around [sync.Map].
+type Map[K, V any] struct {
+	async.Map[K, V]
+}