--- a/internal/async/pool.go
+++ b/internal/async/pool.go
@@ -1,3 +1,5 @@
+//go:build !migrated_fynedo
+
 package async
 
 import "sync"