--- a/internal/async/pool_test.go
+++ b/internal/async/pool_test.go
@@ -18,3 +18,29 @@ func TestPool(t *testing.T) {
 
 	assert.Equal(t, -1, pool.Get())
 }
+
+var sink int
+
+func BenchmarkPool(b *testing.B) {
+	p := &Pool[int]{
+		New: func() int {
+			return 0
+		},
+	}
+
+	b.Run("GetOnly", func(b *testing.B) {
+		b.ReportAllocs()
+		local := 0
+		for i := 0; i < b.N; i++ {
+			local = p.Get()
+		}
+		sink = local
+	})
+
+	b.Run("PutOnly", func(b *testing.B) {
+		b.ReportAllocs()
+		for i := 0; i < b.N; i++ {
+			p.Put(i)
+		}
+	})
+}