<!-- If this is your first pull request for Fyne please read the contributor docs at:
https://github.com/fyne-io/fyne/wiki/Contributing.
Be sure that your work is based off `develop` branch. --> 

### Description:
<!-- A summary of the change included and which issue it addresses.
Please include any relevant motivation and background. -->

The pool was used in places where concurrency actually was important. I solved this by introducing a new `internal/async/migrated` package which has all of the async data structures but with build tags to become single threaded when migrated. This has the added benefit of avoid issues where we accidentally uses the single threaded version although we need safety.

Benchmark results are promising:

Before:
```
BenchmarkPool/GetOnly-22      	11466109	        90.45 ns/op	       0 B/op	       0 allocs/op
BenchmarkPool/PutOnly-22      	28852821	        38.67 ns/op	      35 B/op	       1 allocs/op
```

After:
```
BenchmarkPool/GetOnly-22      	370900233	         3.386 ns/op	       0 B/op	       0 allocs/op
BenchmarkPool/PutOnly-22      	59720872	        16.87 ns/op	       8 B/op	       1 allocs/op
```


### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [ ] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.